requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
bcrypt>=4.0.0
tiktoken>=0.6.0
python-dotenv>=1.0.0
pgvector>=0.2.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
